Fix #155173: Copying large string from string buttons crashes
Button selection may overflow short value range for large strings, this
uses a int instead to avoid crashes when copying from buttons.

Its important to note that buttons currently does not support
handling/showing such large strings in text edit mode (they just hangs
out when trying to get the full string width for selection boxes for
example).

Pull Request: https://projects.blender.org/blender/blender/pulls/155254

+++ source/blender/editors/interface/interface_intern.hh
@@ -204,7 +204,7 @@ struct Button {
   uchar menu_key = 0;
 
   short retval = 0, strwidth = 0, alignnr = 0;
-  short ofs = 0, pos = 0, selsta = 0, selend = 0;
+  int ofs = 0, pos = 0, selsta = 0, selend = 0;
 
   /**
    * Optional color for monochrome icon. Also used as text
